api/requirements.txt
cachelib==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
Flask==2.0.1
Flask-Dance==5.0.0
Flask-Session==0.4.0
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
python-dotenv==0.19.0
requests==2.26.0
requests-oauthlib==1.3.0
SQLAlchemy==1.4.22
urllib3==1.26.6
URLObject==2.4.3
Werkzeug==2.0.1
